Lab10/pyproject.toml
[tool.poetry]
name = "lab10"
version = "0.1.0"
description = "Wątki"
authors = ["Natalia Grzybicka"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
requests = "^2.28.2"
bs4 = "^0.0.1"
Pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
